Add files via upload
My first attempt at using a python code for machine learning. The task consists of creating a prediction system for predicting CO2 emissions based on attributes like engine size, number of cylinders in the vehicle. Along with this, publishing this file is my first attempt of any activity on Github.